added necessary files from my rpg project
I could not figure out how to get an entire project from eclipse to github properly even when using both the video and the internet as a whole. nothing I find is working and the semantics of this project and the few points it'll cost me is not worth the mental breakdown.